requirements.txt
SQLAlchemy==1.3.8
alembic==1.1.0
botocore
flask==1.1.1
flask-sqlalchemy==2.4.0
flask-migrate==2.5.2
zappa
Pillow=6.1.0
psycopg2==2.8.3
requests
slackclient
slackeventsapi
pytest
pytest-flake8
pytest-black
